Adding initial swag runbook (#7)
* Adding initial swag runbook

* First draft of the swag_runbook.md

* Update swag_runbook.md

logo and annual designs are different. the logo does not change every year but we do have new designs every year based on the theme

Co-authored-by: Laura Santamaria <nimbinatus@users.noreply.github.com>

* Remove unneeded line describing important dates

The important dates are below and this line is not needed.

Co-authored-by: Laura Santamaria <nimbinatus@users.noreply.github.com>

* Adding line about holding back some swag for future events

* Updating references from logo to annual swag designs where appropriate

logo - the DevOpsDays brain that stays consistent every year
annual swag designs - the art that is commisioned for t shirts, badges, etc.  This is based on the chosen theme

---------

Co-authored-by: Marisa Sawatphadungkij <marisasawat@gmail.com>
Co-authored-by: Laura Santamaria <nimbinatus@users.noreply.github.com>